+++ src/temp/ProjectViewTabs/ProjectViewTabs.module.scss
@@ -31,8 +31,8 @@ gap: 16px;
     padding-right: 6px;
   }
   .option p {
-    font-size: 12px;
-    font-weight: 700;;
+    display: none;
   }
   
+
